app/requirements.txt
google-cloud==0.34.0
pandas==1.2.3
